chore: move logic inline
