--- src/util/client.js
@@ -1,5 +1,0 @@
-import { version } from '../../package.json';
-
-export const addAlgoliaAgent = client =>
-  client.addAlgoliaAgent &&
-  client.addAlgoliaAgent(`Vue InstantSearch (${version})`);
